Some functions have unit tests in files with different names.
Typically, we like to have a function with a matching _test.c file
containing its unit tests.  Sometimes, functions are so small or so
similar that having multiple test files doesn't make sense.  This
typically happens when we have a case-sensitive and case-insensitive
version of some string function.

So, just add a list of where this happens so we can make sure those
functions are marked as tested.